+++ presentation/src/main/java/com/fernandocejas/android10/sample/presentation/internal/di/components/UserComponent.java
@@ -16,6 +16,7 @@
 package com.fernandocejas.android10.sample.presentation.internal.di.components;
 
 import com.fernandocejas.android10.sample.presentation.internal.di.PerActivity;
+import com.fernandocejas.android10.sample.presentation.internal.di.PerFragment;
 import com.fernandocejas.android10.sample.presentation.internal.di.modules.ActivityModule;
 import com.fernandocejas.android10.sample.presentation.internal.di.modules.UserModule;
 import com.fernandocejas.android10.sample.presentation.view.fragment.UserDetailsFragment;
@@ -26,7 +27,7 @@ import dagger.Component;
  * A scope {@link com.fernandocejas.android10.sample.presentation.internal.di.PerActivity} component.
  * Injects user specific Fragments.
  */
-@PerActivity
+@PerActivity // TODO: this is actually PerFragment
 @Component(dependencies = ApplicationComponent.class, modules = {ActivityModule.class, UserModule.class})
 public interface UserComponent extends ActivityComponent {
   void inject(UserListFragment userListFragment);
